+++ src/routes/pages/calendar/+page.svelte
@@ -7,165 +7,67 @@
     import EventCard from '$lib/components/calendar/event/EventCard.svelte';
     import EventDetails from '$lib/components/calendar/event/EventDetails.svelte';
     import type { Event } from '$lib/components/calendar/event/Event';
-    import { EventType, EventRepeatType } from '$lib/components/calendar/event/Event';
+    import { EventType, EventRepeatType, eventHasCorrectFormat } from '$lib/components/calendar/event/Event';
+    import { isCurrentDay } from '$lib/components/calendar/CalendarFunctions';
 	import { toastController } from 'ionic-svelte';
 	import type { ToastOptions } from '@ionic/core';
 
     let activeDate: Date;
     let eventList: Event[];
     let selectedEvent: Event | null = null;
-    let tmpEvent: Event | null = null;
+    let tmpEvent: Event | undefined;
     let modalOpen: boolean = false;
-    
-    $EventStore = [{ id: 1,
-    title: "Study Group Meeting",
-    slot: 
-        {
-            start: new Date("2024-05-17T10:00:00"),
-            end: new Date("2024-05-17T12:00:00")
-        }
-    ,
-    location: "Library",
-    description: "Discussing upcoming exam topics",
-    professor: "Dr. Smith",
-    type: EventType.CLASS,
-    repeat: EventRepeatType.DAILY,
-    repeatInterval: 2,
-    repeatUntil: new Date("2027-01-01T23:59:59"),
-    notify: true,
-    notifyTime: 30}]
-    
-    $: eventList = $EventStore.filter(item => isCurrentDay(item, activeDate)).sort((a, b) => a.slot.start.getTime() < b.slot.start.getTime() ? -1 : 1);
-    
-    function isCurrentDay(event: Event, active: Date): boolean {
-        const activeDate = new Date(active);
-        const start = new Date(event.slot.start);
-        const end = new Date(event.slot.end? event.slot.end : event.slot.start);
 
-        if(event.repeat == EventRepeatType.NEVER) {
-            return (
-                start.getFullYear() === activeDate.getFullYear() &&
-                start.getMonth() === activeDate.getMonth() &&
-                start.getDate() === activeDate.getDate()
-            );
-        }
-        
-        if (!event.repeatUntil || !event.repeatInterval) return false;
+    {
+    // $EventStore = [{ id: 1,
+    // title: "Study Group Meeting",
+    // slot: 
+    //     {
+    //         start: new Date("2024-05-17T10:00:00"),
+    //         end: new Date("2024-05-17T12:00:00")
+    //     }
+    // ,
+    // location: "Library",
+    // description: "Discussing upcoming exam topics",
+    // professor: "Dr. Smith",
+    // type: EventType.CLASS,
+    // repeat: EventRepeatType.DAILY,
+    // repeatInterval: 2,
+    // repeatUntil: new Date("2027-01-01T23:59:59"),
+    // notify: true,
+    // notifyTime: 30}];
 
-        const repeatUntil = new Date(event.repeatUntil);
-        const repeatInterval = event.repeatInterval;
-        
-        let distance: number = 0;
-
-        if(event.repeat == EventRepeatType.DAILY) {
-            const intervalMilliseconds = repeatInterval * 24 * 60 * 60 * 1000;
-            const totalDuration = repeatUntil.getTime() - start.getTime();
-            const fullIntervals = Math.floor(totalDuration / intervalMilliseconds);
-            const lastOccurrence = new Date(start.getTime() + fullIntervals * intervalMilliseconds);
-
-            const dayDistance = (date1: Date, date2: Date) => {
-                const diffTime = date2.getTime() - date1.getTime();
-                return Math.ceil(diffTime / (1000 * 60 * 60 * 24));
-            };
-
-            const distance = dayDistance(start, activeDate);
-            return (
-                distance >= 0 &&
-                distance % repeatInterval === 0 &&
-                activeDate.getTime() <= lastOccurrence.getTime()
-            );
-        }
-
-        if (event.repeat == EventRepeatType.WEEKLY) {
-            const dayDistance = (date1: Date, date2: Date): number => {
-                return (new Date(date1.setHours(0,0,0,0)).getTime() - new Date(date2.setHours(0,0,0,0)).getTime()) / (1000 * 60 * 60 * 24);
-            }
-
-            const intervalMilliseconds = repeatInterval * 7 * 24 * 60 * 60 * 1000;
-            const totalDuration = repeatUntil.getTime() - start.getTime();
-            const fullIntervals = Math.floor(totalDuration / intervalMilliseconds);
-            const lastOccurrence = new Date(start.getTime() + fullIntervals * intervalMilliseconds);
-            
-            distance = dayDistance(activeDate, start);
-            return (distance >= 0 && distance % (repeatInterval * 7) === 0 && activeDate.getTime() <= lastOccurrence.getTime() && start.getDay() === activeDate.getDay());
-        }
-
-        if (event.repeat == EventRepeatType.MONTHLY) {
-            const getMonthsDifference = (startDate: Date, endDate: Date) => {
-                return (endDate.getFullYear() - startDate.getFullYear()) * 12 + (endDate.getMonth() - startDate.getMonth());
-            };
-
-            const totalMonths = getMonthsDifference(start, repeatUntil);
-            const fullIntervals = Math.floor(totalMonths / repeatInterval);
-            const lastOccurrence = new Date(start);
-            lastOccurrence.setMonth(start.getMonth() + fullIntervals * repeatInterval);
-
-            const isSameDayOfMonth = (date1: Date, date2: Date) => date1.getDate() === date2.getDate();
-
-            const monthDistance = getMonthsDifference(start, activeDate);
-            return (
-                monthDistance >= 0 &&
-                monthDistance % repeatInterval === 0 &&
-                activeDate.getTime() <= lastOccurrence.getTime() &&
-                isSameDayOfMonth(start, activeDate)
-            );
-        }
-
-        if(event.repeat == EventRepeatType.YEARLY) {
-            const getYearsDifference = (startDate: Date, endDate: Date) => {
-                return endDate.getFullYear() - startDate.getFullYear();
-            };
-
-            const totalYears = getYearsDifference(start, repeatUntil);
-            const fullIntervals = Math.floor(totalYears / repeatInterval);
-            const lastOccurrence = new Date(start);
-            lastOccurrence.setFullYear(start.getFullYear() + fullIntervals * repeatInterval);
-
-            const isSameDayOfYear = (date1: Date, date2: Date) => {
-                return date1.getDate() === date2.getDate() && date1.getMonth() === date2.getMonth();
-            };
-
-            const yearDistance = getYearsDifference(start, activeDate);
-            return (
-                yearDistance >= 0 &&
-                yearDistance % repeatInterval === 0 &&
-                activeDate.getTime() <= lastOccurrence.getTime() &&
-                isSameDayOfYear(start, activeDate)
-            );
-        }
-
-        return false;
+    // To clear the EventStore, uncomment the line below
+    // $EventStore = [];
     }
 
+    $: eventList = $EventStore.filter(item => isCurrentDay(item, activeDate)).sort((a, b) => a.slot.start.getTime() < b.slot.start.getTime() ? -1 : 1);
+    $: console.log(eventList);
+    
     function sumbit() {
-        if(tmpEvent?.id == selectedEvent?.id) {
-            console.log(tmpEvent?.repeatUntil);
-            if(!eventHasCorrectFormat(tmpEvent)) {
-                showToast({
-						color: 'danger',
-						duration: 3000,
-						message: 'Τσέκαρε τα στοιχεία του συμβάντος!',
-						mode: 'ios',
-						translucent: true,
-						layout: 'stacked',
-						positionAnchor: "bottom",
-						cssClass: 'custom-toast'
-					});
-                return;
-            }            
-            const index = $EventStore.findIndex(x => x.id == tmpEvent?.id);
-            
-            if(index != -1 && tmpEvent!=null) {
-                $EventStore[index] = tmpEvent;
-            } else if (tmpEvent!=null) {
-                $EventStore = [...$EventStore, tmpEvent];
-            }
-            selectedEvent = null;
-            tmpEvent = null;
-            modalOpen = false;
-        } else {
-            modalOpen=false;
+        if(!eventHasCorrectFormat(tmpEvent)) {
+            showToast({
+                    color: 'danger',
+                    duration: 3000,
+                    message: 'Τσέκαρε τα στοιχεία του συμβάντος!',
+                    mode: 'ios',
+                    translucent: true,
+                    layout: 'stacked',
+                    positionAnchor: "bottom",
+                    cssClass: 'custom-toast'
+                });
+            return;
+        }            
+        const index = $EventStore.findIndex(x => x.id == tmpEvent?.id);
+        console.log(tmpEvent, index);
+        if(index != -1 && tmpEvent != undefined) {
+            $EventStore[index] = tmpEvent;
+        } else if (tmpEvent != undefined) {
+            $EventStore = [...$EventStore, tmpEvent];
         }
+        selectedEvent = null;
+        tmpEvent = undefined;
+        modalOpen = false;
     }    
 
 	async function showToast(toast: ToastOptions){
@@ -173,37 +75,15 @@
 		toast_.present();
 	}
 
-    function eventHasCorrectFormat(event: Event | null): boolean {
-        return (event!=null && event.title != undefined && event.title.length > 0 && new Date(event.slot.start).getTime() <= new Date(event.slot.end).getTime());
-    }
-
     function setupModal() {
-        if(selectedEvent == null) {
-            if(!tmpEvent) {
-                tmpEvent = {
-                    id: new Date().getTime(),
-                    title: "",
-                    slot: {
-                        start: new Date(),
-                        end: new Date(new Date().getTime() + 3600000)
-                    },
-                    type: EventType.TASK,
-                    location: "",
-                    description: "",
-                    professor: "",
-                    notifyTime: 30,
-                    repeat: EventRepeatType.NEVER,
-                    notify: false
-                }
-            }
-        } else {
+        if(selectedEvent !== null) {
             tmpEvent = JSON.parse(JSON.stringify(selectedEvent));
             if (tmpEvent) {
                 tmpEvent.slot.start = new Date(selectedEvent?.slot.start);
                 tmpEvent.slot.end = new Date(selectedEvent?.slot.end);
             }
         }
-        modalOpen=true;
+        modalOpen = true;
     }
 
 </script>
@@ -265,7 +145,7 @@
                     </ion-button>
                 </ion-buttons>
             </ion-toolbar>
-                <EventDetails bind:copyEvent={tmpEvent} />
+            <EventDetails bind:copyEvent={tmpEvent} />
         </ion-modal>
     </ion-content>
 </ion-tab>
